interview/src/test/java/apple/_767_Reorganize_String/SolutionITest.java
package apple._767_Reorganize_String;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static definition._767_Reorganize_String.Assertions.assertValid;
import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 2/12/2021
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase1(Solution solution) {
    final String s = "aaabbcc", actual = solution.reorganizeString(s);
    assertValid(s, actual);
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase2(Solution solution) {
    final String s = "aaab", actual = solution.reorganizeString(s);
    assertEquals("", solution.reorganizeString(actual));
  }

  static Stream<Solution> solutionProvider() {
    return Stream.of(
        new SolutionI()
    );
  }
}